[logdog] Refactor archivist to form a pipeline.
Currently the archivists consume a batch (~500 tasks) from all
available tasks. The VAST majority of tasks are very tiny logs. This
means that archivists which consume a batch of 499 tiny logs and 1 big
log will end up doing almost nothing while the 1 big log completes.

This CL refactors the archivist into a pipeline consisting of three phases:
  * A leaser goroutine
  * A job executor dispatcher
  * An ACKer dispatcher

The leaser feeds leased tasks into the job executor channel; The job
executor can buffer up to 2x the archivist's max concurrency. When
there are fewer jobs than that, the leaser will lease the next batch
of work.

The ACKer goroutine buffer ACKs until they're a group of 500, or until
they're 10 minutes old, whichever is first. The ACKing process has a
substantial buffer to prevent it from blocking work, since ACKs are
really an optimization in the grand scheme of things; an ACKable task
is already recorded in the datastore as being archived and the
coordinator will not re-issue it.

Unlike previously, the concurrency setting for the archivers will now
actually be relevant ~100% of the time the archiver is working, because
the workers should never starve.

R=tandrii@chromium.org, vadimsh@chromium.org

Bug: 1061187
Change-Id: I07b25a566f5479945f50db55a796bfc20c82b538
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2121368
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>