[logdog] Deepen jobChan buffer to hide appengine RPC latency.
Currently it looks like the appengine handler takes close to 60s
to return a batch, and it takes ~3 minutes between RPCs. This means
that approximately 1 minute out of every 3, the archivist is just
chilling.

The numbers here are based on the current monitored values rounded
in pessimistic directions. This will compute to 288 (up from 32) for
the current configuration.

R=vadimsh@google.com, yiwzhang@google.com

Bug: 1061187
Change-Id: I7d8d12cb7316d11f1839d5fc4e498bfc95c2ce74