[gce-provider] Add binary_config on config in datastore and change related read/write
Only for binary proto changes on Config. Project and VM changes will be in other CLs.
It's the code implementation for Step 1 in this doc- https://docs.google.com/document/d/1i2qQXH861ifhEPRsN6uaT3yacrSLF0znu6K3J4cSQao/edit?usp=sharing

Overral workflow:

Before every datastore.put call, It will copy the text config to the binary config field.
Because both text and binary protos are up-to-date, we can directly use text protos in all read flows.
So that we avoid the extra work to change the code in reading flow to use binary protos info and change code back later.

The general idea is a sightly different from that doc (which said writing only goes to binary and reading from binary then fall backs to text if nil).
Because I found the new way is more nature, organized and requires less code changes.

Basically, the binary proto field is like the mirror or a standby for text field.
After Step 2 (in which we migrated all text to binary), we can change the `gae:"config,noindex"` to `gae:"binary_config,noindex"`.

Bug: 1036145
Change-Id: I2b114c86d2ed3af6eddfba3af31bd5bba5f0b604