[resultdb] Add Invocation.producer_resource
When test results look suspicious, a user may want to go to the build/task
that has produced the results, to see if anything went wrong there.
For that, UI needs to provide a link.
For that, we need a generic way to refer to the task/build that has
produced the results.

Add a canonical way to annotate invocations with a resource that
produced the results, e.g. a Swarming task or a Buildbucket build.
Use https://aip.dev/122#full-resource-names as the format, because
UI can easily produce a URL from that by prepending "https".

Rename CIA group from "luci-resultdb-custom-invocation-id"
to "luci-resultdb-trusted-invocation-creators" to account for the case.

Before deployment, we need to add a new Spanner column and rename
the CIA group.

Change-Id: I4d8047dd65e27b3df35e9b52cf527a4394aa42f1