[resultdb] Add Invocation.producer_resource
Add a canonical way to annotate invocations with a resource that
produced the results, e.g. a Swarming task or a Buildbucket build.

Rename CIA group from "luci-resultdb-custom-invocation-id"
to "luci-resultdb-trusted-invocation-creators" to account for the case.

Before deployment, we need to add a new Spanner column and rename
the CIA group.

Change-Id: I4d8047dd65e27b3df35e9b52cf527a4394aa42f1