Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> Hey Weiwei, can you explain a bit to me about what problem this solves? Can you give an example in the results UI where we want to use this?

This is the equivalent of react context.
https://reactjs.org/docs/context.html

It's basically dependency injection for lit-element. So instead of passing a parameter all the way down through component properties, the parent element can simply inject it to the context.

The most important usage of this is to share app-level states without using a global variable.
In ResultUI, for example, selectedTest, resultDdPrpcClient can be injected through context.

Some benifits:
1. Avoid passing parameters all the way down. Make the code more concise and avoid triggering unnecessary updates in parent nodes.
2. Avoid using a global variant, making it easier to test the component.

Patch-set: 7