Update patch set 2
Patch Set 2:

When the version will be regenerated? How to decide when to regenerate it? Where will it end up? Will regenerating this version number trigger rebuild of everything?

I'm not a fan of using a single version string for everything. luci-go is a collection of utilities many of which are totally independent and it makes little sense to version them together.

We had relatively good experience using semantic versioning (<major>.<minor>.<rev>) for individual tools we release as CIPD packages, for example:
https://chromium.googlesource.com/infra/luci/luci-go/+/refs/heads/master/cipd/client/cipd/client.go#134
https://chromium.googlesource.com/infra/luci/luci-go/+/refs/heads/master/lucicfg/version.go#27
https://chromium.googlesource.com/infra/infra/+/refs/heads/master/go/src/infra/cmd/cloudbuildhelper/main.go#27

Rules we used for updating semantic versions:
1. If the tool has been rewritten significantly or learned some new fundamental feature, bump the major version. This happens rarely. (It happened when CIPD client switched from Cloud Endpoints protocol to pRPC protocol).
2. If there's some backward incompatible change that potentially requires clients migration or clients may wish to do checks like "if <versionX> { do X } els { do Y}", bump the minor version.
3. If the tool is updated with a new feature in a backward compatible way and you want to release this change to the clients, bump the <rev> component of the version.
4. If you do not plan to release the change right away and want to add more stuff on top of it, don't touch the version number (otherwise we'd need to update it every commit which is not very convenient).

Patch-set: 2
Reviewer: Gerrit User 1002539 <1002539@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>