[swarming] Add -eager mode to collect command
If `swarming collect` is called with the `-eager` flag, it will return
"eagerly" as soon as *any* task has completed, instead of waiting until
all tasks have finished.

This will speed up collecting and processing of many tasks. We'll be
able to process each task as soon as it finishes, so we'll generally
only process one task at a time instead of waiting until *all* tasks
finish before processing them all sequentially.

I also added logic to ensure that if other tasks complete in the time it
takes to download the first task's outputs, then we'll wait to download
those other tasks' outputs before exiting. This may not be necessary (we
don't wait for downloading to complete if we hit the collect timeout)
but it seemed nice to have. It does add a lot more complexity to the
code, so I can remove it if preferred.

Bug: fuchsia:47053
Change-Id: Ifa71d2effa18e07b04f2488da0262bfeb304db64