[resultdb] update flags for result_sink_wrapper and init Recorder client
(1) Replaced Flags.invocationIDFile with Flags.invocation because the Swarming task will generate and provide the invocation.
(2) Removed Flags.resultFiles, as it's not clear how it will be used.
It was proposed that sinkServer supports uploading test results from a file, but the feature is not currently supported.
This flag can be added back later if necessary.
(3) Added Flags.updateToken

Change-Id: If9f3ec346c3b71f7bddf758c36dbe0b9ad5f342e
Bug: 1017288
R=nodir@google.com