[resultdb] Remove an assumption
Currently `rdb call` assumes that the invocation in the request is the
same as the invocation in LUCI_CONTEXT because it implicitly sends
the update token from LUCI_CONTEXT. Remove the assumption by making
it explicitly: add -include-update-token flag.

Fail if -include-update-token is set, but there is no update token.

Change-Id: I08968dab414155d6ad9cc65ec1cc9f52028efdbd