Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Hm, I thought the "not connected" case was handled? Doesn't the conversion run as the first thing in "manage bot"? It should happen even if the bot isn't connected, shouldn't it?
> 
> Yes, you are right. It's been handled. You can find the caller code here - https://source.chromium.org/chromium/infra/infra/+/master:go/src/go.chromium.org/luci/gce/appengine/backend/bots.go;l=203?q=convertToBinary
> 
> I saw a few entities don't have binary format under the created but unconnected status. That's why I got previous conclusion. But I queried those again after some time. They all have binary. It's due to the latency. The manage-bot async job hadn't picked those up at that time. 
> 
> Considering the above situation only happens for newly created VMs, maybe the deploying strategy can be as follows.
> 1. make sure no records for (VMs `create=0`)
> 2. wait for all those (VMs `connect=0`) get populated with binary 
> 
> Does it make sense? Maybe we can only find those perfect timing at night.

Checking for such records and making the code switch on the server won't be atomic. It's difficult to check because VM entities are deleted and recreated, so you may see that there aren't any with create=0 but a moment later a new entity is written to the datastore.

The alternative is to write another CL which is deployed before this, in which the binary config is set during VM entity creation. Then we still have to check to make sure there are no VMs with create=0, but at least we'll know that if a new VM is created it'll have binary config populated.

Patch-set: 4
Reviewer: Gerrit User 1003211 <1003211@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>