Update patch set 9
Abandoned

Since the issue this CL trying to fix is a rare occurrence (can only happen if user select the last branch in the test nav tree and the tests are not fully loaded), and does not break the page in any major way. It's probably not worth adding this level of complexity to the code.

Patch-set: 9
Status: abandoned
Tag: autogenerated:gerrit:abandon