[resultsdb] sink.Run no longer closes the connection
(1) Run no longer closes the server. The owner of the sinkServer instance,
which is whomever called NewServer(), is responsible to invoke Close() and
wait for all the outgoing requests processed.

(2) sink no longer relies on errC for internal communications
an error on s.errC is consumable, and, it's racy that s.Start() and s.Run()
wait for a notification on the channel.
Also, the caller of NewServer() can wait for an event on the channel with ErrC,
even while Run() is still being executed.

Done has been added to indicate that the server has stopped and processed
all the pending sink requests, and sink itself no longer relies on errC to
determine if the server has stopped.

Change-Id: I2dfd2296d5ceae6c5b88ed50b2423d77ebe7a458