[resultsdb] sink.Run doesn't close the server
(1) Run no longer closes the server. Whoever called NewServer() is responsible to invoke Close() and wait for the server to process all the ongoing requests.

(2) sink no longer relies on errC to detect the termination of the goroutine with httpsrv.
an error on s.errC is consumable, but all s.Start(), s.Run(), and s.the caller of Errc() can race each other to catch the signal.
Done has been added to resolve the issue. It is closed when the server has successfully stoppsed and processed all the ongoing requests.
The caller of Run() or Start() can still be able to obtain the error from ErrC, when Done is closed.

Change-Id: I2dfd2296d5ceae6c5b88ed50b2423d77ebe7a458
R=nodir@chromium.org
Bug:1017288