[resultsdb] sink.Run no longer closes the connection
(1) Run no longer closes the server. Whoever called NewServer() is responsible to invoke Close() and wait for all the outgoing requests processed.

(2) sink no longer relies on errC for internal communications
- an error on s.errC is consumable, but
- s.Start(), s.Run(), and the caller of s.ErrC() can race each other to catch the signal.

Done has been added to indicate that the server has stopped and processed all the pending sink requests, and sink itself no longer relies on errC to determine if the server has stopped.

The caller of Run() or Start() can extract an error from ErrC, when Done is closed.


Change-Id: I2dfd2296d5ceae6c5b88ed50b2423d77ebe7a458