[resultui] implement test selection
Allow users to select a branch a test from the navigation panel.

Know issue:
When the user select a node that is not fully loaded, new tests might be
loaded to the tree, causing the tree to expand and/or the selected node
to be unelided.
This should be a rare occurrence since
1. only the last leaf and its ancestors can be not fully loaded.
2. when VARIANTS_WITH_UNEXPECTED_RESULTS filter is used, the number of
tests should be low and all nodes should be fully loaded on first page.


Bug:1044351
Change-Id: I115fb6a2ebad3762c46097155626297ba842429e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2135086
Auto-Submit: Weiwei Lin <weiweilin@google.com>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Joey Scarr <jsca@chromium.org>
Commit-Queue: Weiwei Lin <weiweilin@google.com>