[luci-auth] Make `luci-auth context` fail if there's no credentials.
Previously it would initiate a login flow. `luci-auth context` is now
used on bots and it is important they never block on input.

To setup local credentials for the context, humans now need to run
$ luci-auth login -context

It is just a shortcut for "luci-auth login -scopes <giant list of
scopes needed by context>".

R=iannucci@chromium.org

Change-Id: I18a03382df67e01acac28597c9e9bd546c9e2325