[resultdb] Disallow ^ and $ in regexp.
To implement test result predicate on top of Artifacts.ParentId column,
we need to be able to prepend "tr/" to the user-supplied pattern.
To do that correctly, the user-supplied pattern must not have ^.

Disallow $ and ^ in user-supplied regexpes. Also it is educational:
without this the user might not know that this is actually not necessary
because we do it automatically.

Bug: 1071258
Change-Id: I8a05f8d50855dbffb30457f287c3bcbb27814f80