Update patch set 3
Patch Set 3:

it occurred to me that this also does not allow \$ in the end. On one hand, this is very unlikely, but on the other hand we could use https://pkg.go.dev/regexp/syntax?tab=doc#Parse
Perhaps we should use it if someone reports a problem with this.

Patch-set: 3