Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> > - paint (expectedly) "passed" green?
> I considered this when designing the old version. But since there could be expectedly failed tests, wouldn't it be weird to color "failed" green? Likewise, I'm not sure about coloring the status red when the result is unexpected. Because that means "unexpectedly passed" will be red too, which feels weird.

Yes, it does feel weird, but this is not new and Chrome developers are accustomed to this. Moreover, painting "unexpected pass" green or not-red will not grab user's attention, although it should because it is a reason to reject the CL. Users do want to see those. Ultimately, this is not something we can change, so i think we should just play by the existing rules.

For some background: Chrome runs https://web-platform-tests.org/ which is a battery of tests for *all* browsers. It is developed in a separate repo and being imported into chrome. When a test for *new* feature is imported, the test is *expected to pass*. Then, engineers implement the feature and eventually the test passes. In the CL where the test finally starts to pass, the engineer MUST flip the bit and expect a pass. If they don't do that, then we'd end up with incorrect expectations: the test is passing, but it is expected to fail: that's why the engineer must flip the bit in the same CL, and we must reject CLs that don't do that.

Patch-set: 1
Reviewer: Gerrit User 1114299 <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>