[resultui] fix context lib breaks when applying the mixin multiple times
Originally, the contextConsumer/Provider mixin can not be applied
multiple times, because the mixins add private properties to the class.
This CL convert those private properties to a variables stored in the
closure. So the mixin can be applied multiple times.

Applying mixin multiple times is useful when you want to consume contexts
from different providers.


Change-Id: I19d5f965d108c9b850d22b9cf4d5ce651f5ab786