Update patch set 6
Patch Set 6:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > > Patch Set 5: Code-Review+1
> > > 
> > > LGTM. Got a demo?
> > Oops. Forgot that.
> > https://weiweilin-dot-resultui-dot-luci-milo-dev.appspot.com/inv/task%3Achromium-swarm_appspot_com%3A4b9997b3c6e51211/test-results
> 
> Interesting. So if you leave out the /test-results suffix, we get page not found. Can we just have it load the page and default to the test-results tab?

That was a bug. Fixed.

> 
> Also, I feel like it's more common to use a fragment for something like this (#test-results rather than /test-results) - is there a reason we can't do that?

We can. But the router I used doesn't support routing base on fragments.
We can simply implement our own conditional rendering. But that means we can't take advantage the code splitting features built in to the router lib.
While this is less of an issue now, since the other tab, invocation_details, is pretty small. Once we add more complex tabs, user will be forced to load a lot of js on the first load.

Patch-set: 6