[resultdb] add sink.Done and sink.Shutdown
(1) sink no longer relies on errC to detect the termination of
the goroutine with httpsrv.

An error on s.errC is consumable, and all of s.Start(), s.Run(), and
the caller of ErrC() can race each other to catch the signal. Done has
been added to resolve the issue. It is closed when the server has
successfully stoppsed and processed all the ongoing requests.

(2) Cancelling the context, passed to Start(), no longer results in
the server terminated.

Closing the server context doesn't halt the channel workers, and causes
sinkServer.CloseSinkServer() returns immediately without waiting for
the channel to be drained, Close() or Shutdown() should be used to
close the server, instead.

Change-Id: I2dfd2296d5ceae6c5b88ed50b2423d77ebe7a458
R=nodir@chromium.org
Bug:1017288
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2160002
Commit-Queue: Scott Lee <ddoman@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>