[resultdb] Run is a global function under package sink
The intent of this change is to protect Run from misuse of other functions
under sink. If sink.Run() is used, there is no need to invoke Start(), Close(),
Shutdown(), Done(), and Err().

If a user of sink.Run wants to terminate Run in the middle of the execution,
the user should just send a signal to the callback to voluntarily finish.

Change-Id: I51789ccc7e94caee1943bc9059a9d3dc94b2d6d9
Bug:1017288
R=nodir@chromium.org