Update patch set 3
Patch Set 3:

The UI looks awesome. Dragging is great.

UX experience: it is not obvious that I need to double-click to expand/collapse. The cursor is a hand and normally this means on click is enough. As a user, I might not discover the double-click and perceive this as a bug. It is especially confusing when it is collapsed: dragging is not supported anyway. Why not just one click? I think we can distinguish sole click and dragging, because when dragging the mouse button is held for a prolonged period.

Patch-set: 3
Reviewer: Gerrit User 1114299 <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>