[resultdb] Run is a global function under package sink
The intent of this change is to protect Run from misuse of other sink
functions. If sink.Run() is used, there is no need to invoke
Start(), Close(), Shutdown(), Done(), and Err().

If a user of sink.Run wants to terminate Run in the middle of
the execution, the user can just send a signal to the callback
so that the callback voluntarily finishes its execution.

Change-Id: I51789ccc7e94caee1943bc9059a9d3dc94b2d6d9
Bug:1017288
R=nodir@chromium.org
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2174218
Commit-Queue: Scott Lee <ddoman@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>