[resultdb] Optimize leasing
Currently leasing of a claimed tasks takes 1-10s which is very long.
One theory is that some time is spent on fetching the payload and
on extra roundtrip.

Move the "is it leased" check to Spanner. This way we don't fetch
payload of a leased task, and this is 1 RPC fewer.

Change-Id: I59db7f8beb38c507cecd7cea0e256140eb0b6334
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2176526
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>