[milo] Remove releasing commands from Makefile.
I've left the up-dev-* commands and converted them so that they push to
a custom version, which is the user's username. I've also left in
up-resultui since that's not pushed automatically yet.

I've also added instructions to the README for the new way to release
Milo, and deleted the old README since the entire thing was obsolete.

Change-Id: Idabf1f9ada98edd43decc40777ef61085d526fcf
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2172287
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Joey Scarr <jsca@chromium.org>