[logdog] Fix that the expired stream is not drained when flushing
The issue is that when bundler is flushing, it checks whether the stream
snapshot is empty to decide whether flushing is done. However, the
absence of a stream in the stream snapshot does not indicate that this
stream has been drained after change: http://crrev.com/c/2176410 .
Instead, we should look at the stream registry in bundler. This change
also makes sure if a stream is drained, it will be unregistered from the
bundler.

Failed build: https://ci.chromium.org/p/infra-internal/builders/prod/git-cache-chromium-host/b8881298048127898592

R=iannucci

Bug: 1077149
Change-Id: Ib52f4bd176e5e278e85e12aa94bcd3b39cabf2e9