[resultdb] Restore -task-workers
Turns out -task-workers is needed, because its value depends on the
size of tasks we process. On staging and dev instances, we process
small invocations - the task invocations. On prod, we process large,
build-level invocations, that combine tens of thousands of results.
With such large invocations, it is better have smaller -task-workers.

Long term, we should replace ResultDB's task processing with a more
sophisitcated solution, for example a combination of Spanner table and
PubSub queue. But for now, restore -task-workers flag.

Change-Id: Ieb8c5045d6fbf2e13dd3f13d1dfec43ecf7965f3