[resultdb] Expose host metadata
The host that ResultDB needs to redirect to, depends on the host of the
incoming request. This is because ResultDB can be accessed via
stable and canary host. It needs to be able to tell if is is currently
being accessed as stable or canary. Expose "host" metadata key, based
on http.Request.Host.

Bug: 1077057
Change-Id: I5c90b64551b0c5d45a14ba0c9707119e871f21f8