cq: validate location_regexp[_exclude] of triggered_by relationships.
It shouldn't be possible for a CL to match a triggered builder, but not
its parent. The new checks are sufficient, but not strictly necessary,
that is some valid configs will be rejected. This is fine for now.

This change is not backwards compatible, but until last week no
location_regexp[_exclude] were even allowed with triggered_by builders.

R=machenbach, qyearsley

Bug: 1074004
Change-Id: If1a9f8af6a345eb6dbcc1110b0f061fe026ea25d