[buildbucket] Proxy GetBuild requests to the Python service
Since dispatch.yaml is global configuration, splitting a portion of
traffic over to the Go service is impossible. Instead, whenever the Go
service receives a request, proxy requests to the Python service and
return its response instead. This allows switching over fully to this
service safely. In subsequent CLs, this code will be modified to allow a
portion of requests through to the Go service, achieving a programmatic
traffic split. Once 100% of traffic is served by Go, remove the proxy.

Bug: 1042991
Change-Id: I00dffab6322d4921f6e66cc85a55983364475cef