Update TreeCloser status even if the status is the same.
The previous logic was incorrect. Since we compare TreeCloser timestamp
to the current status timestamp, we do in fact want to update even if
the status is the same.

For example, perhaps a build fails and causes the tree to be closed, the
sheriff manually reopens, and then the same builder fails again. In this
case we want to close the tree, and hence we need to update the
timestamp so that it's newer than that of the status update.

Also, moved the SUCCESS || FAILURE check earlier, since it doesn't
depend on the details of any particular TreeCloser.

Bug: 1051371
Change-Id: I3a4e297ef61c8b81c14740847d5f2599c1bbbbb8