Fix a bug in deleting dead tree closers.
There were two problems in the existing code. Firstly, the format of the
elements of liveTreeClosers wasn't correct. Secondly, removeDescendants
would only check the ID of the entity itself. This is fine for builders
and templates, which have a unique ID within the project. But TreeCloser
IDs are only unique in combination with their parent key.

Bug: 1051371
Change-Id: If9bccf0b6f82ea8e82d3558fdc41758d70111664