Update patch set 4
Patch Set 4:

> Patch Set 2: Code-Review+1
> 
> (1 comment)
> 
> LGTM
> 
> high-level question: is this main.go temporary and we plan to embed this new UI into Milo?

Somehow missed this.
I think we will keep this main.go. Because this go server doesn't have much to do with the milo one anyway.
And the new UI will probably be added to milo as a link, unless we have a way to share the auth token between the server and the browser.
Otherwise, it would be weird to login twice on the same page.

Patch-set: 4