[swarming] Use SwarmingService for bots command
The logic in bots.go to create a swarming client duplicated logic in
common.go to create a SwarmingService. To deduplicate, I had bots.go
use a SwarmingService and moved some RPC logic into common.go.

Also, fix a bug where an empty list of bots was serialized to null
instead of an empty JSON list.

Change-Id: I5950f18da4707073db181cbc89dc4c3831bb378e