[swarming] Use SwarmingService for tasks command
The logic in tasks.go to create a swarming client duplicated logic in
common.go to create a SwarmingService. To deduplicate, I had tasks.go
use a SwarmingService and moved some RPC logic into common.go.

Also, fix a bug where an empty list of tasks was serialized to null
instead of an empty JSON list.

Change-Id: If79c1b470dd32f8091796edbec381a1afe4190be