[swarming] Set a 90s timeout on TaskOutput request
This should resolve hangs when downloading task stdout. maruel@
suggested 90 seconds because it's a reasonable buffer beyond the
60-second deadline that the Swarming server sets internally for
responding to requests.

We only set the timeout on the TaskOutput request for now to make sure
that it fixes the issue. If it does, we'll set the timeout on the
http.Client used by the swarming service so that the timeout is applied
to all requests.

Bug: fuchsia:50776
Change-Id: I8b8ae02d69be943cd51b13fabfca347472eea057
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2202219
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>