Allow lists of regexes.
The equivalent of these fields in Gatekeeper are most commonly used to
supply a list of step names. Manually writing out something like:
  step1|step2|step3

is harder to read and diffs worse than having a list:
  step1,
  step1,
  step3,

So in this CL we accept the latter, and convert it into the former
during validation.

Bug: 1051371
Change-Id: I521ec73e2fe5c00d95bb8aeaccf8d803494dd790
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2198301
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Auto-Submit: Owen Rodley <orodley@chromium.org>