[tokenserver] Add protos and boilerplate for MintServiceAccountToken.
It is a replacement for current MintOAuthTokenGrant and
MintOAuthTokenViaGrant. Plus I decided to add a support for
ID tokens right away, since it seems we'll need them soon too.

The new implementation will share almost no code with the old
one (and will be MUCH simpler), so it is more convenient to set it
up as a completely separate Token Server "subsystem".

Naming is hard though, so for now the package is called
"serviceaccountsv2". Once the old implementation is gone, we can
rename the new implementation to "serviceaccounts".

R=iannucci@chromium.org, tandrii@chromium.org
BUG=1082960

Change-Id: I96667e9e6ffab03a56a75447ee47f786a991f0ed
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2202472
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>