[authtest] Implement mocking of realm permissions in FakeDB.
Also cleanup authtest package a little bit and fix discovered
cases of its misusage.

This is not a trivial change because FakeDB used to be just a map
in disguise (ID => its groups), but now we need to store differently
structured data in it.

R=tandrii@chromium.org, nodir@chromium.org
BUG=1070761

Change-Id: I50cc6a7fd052db06aeaa61d4891ffbaa387122a9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2208532
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>