[swarming] Set timeout on all Swarming RPCs
https://crrev.com/c/2202219, which added a 90-second timeout on the
TaskOutput request, has fixed the hangs we were seeing on that endpoint.
There haven't been any hangs in the last 4 days, and previously we had
at least a handful of hangs every day:
http://plx/scripts2/swarming-collect-hangs

maruel@ indicated that this issue is likely to affect all Swarming
endpoints, so it makes sense to set a timeout on all requests.

Change-Id: Icc4c4e6926ada730b4459ce136157df093d82db1
Bug: fuchsia:50776
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2208539
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Oliver Newman <olivernewman@google.com>