Fix another issue with checking the status timestamp.
The idea behind only paying attention to new builds is to avoid closing
the tree after a sheriff has manually opened it. In this case there's
still a failing builder, but we ignore it since it failed before the
last update to the tree.

However, this causes us to incorrectly *open* the tree in some cases.
For example, builder A fails, we close the tree, then builder B passes.
If we only look at new builds, we'll only pay attention to builder B,
and conclude that since all the builders are passing, we can open the
tree.

This CL fixes that by adding an additional check that *all* builders are
passing before opening the tree, not just those builders that have had
new builds.

Previously:
* Any new failures? Close the tree.
* Otherwise, open the tree.

Now:
* Any new failures? Close the tree.
* All builders passing? Open the tree.
* Otherwise, don't do anything.
Bug: 1051371
Change-Id: I4a46850cfc1e3e2185a0efa324c6f6730d5921e1