[lucicfg] Update generation of Executables to no longer emit Recipe.
The Recipe field is deprecated, and is just a shortcut for setting up
an executable and one hard-coded property ("recipe"). Do that here
in lucicfg so we can remove it from buildbucket eventually.

Also this is necessary to allow opt-in for luciexe, because buildbucket
currently treats the 'recipe' field and 'exe' fields as mutually
exclusive.

R=vadimsh@chromium.org, yiwzhang@google.com

Bug: 1015181
Change-Id: I07c052a3f5aa12887cbc37b645c21493cf0d69b3