[server] Add a statsd server that can convert metrics to tsmon.
statsd is a simple UDP-based protocol for receiving metric pushes
from (primarily) localhost processes. Each UDP packet is (usually)
a single line like "something.blah.metric:123|c", where the
dot-delimited string encode both a metric name and its fields.

This server will be used to ingest metrics from Envoy and transform
them into something tsmon understands. The transformation rules are
defined via a config file. See docker/envoy/statsd-to-tsmon.cfg for
an example (it will be extended in the next CL).

The primary design principle is to be as lightweight as possible.
Envoy sends several thousand metrics per minute plus a bunch of
metrics per **every single request** it handles. Most of them we are
not interested it, and should skip as fast as possible. Additionally,
this server will be running as a side car container in Envoy pods,
it can't afford to steal Envoy's CPU resources.

R=ddoman@chromium.org

Change-Id: I2796c3e20e2e295e6ac28d479f37863d0d2722e0