[milo] Merge `resultui` module in `default`.
Using magic of symlinks: all "web" stuff is kept in resultui/ exactly
as is. Only files necessary for GAE deployment are symlinked into
appropriate places in appengine/.

Note that `/inv/.*` regexp in app.yaml may need to be updated to
capture all routes that are supposed to be served by the ResultUI
frontend code.

R=weiweilin@google.com, jsca@chromium.org
BUG=1073318

Change-Id: I8ec0ac52600c8c897749087f73175ac3ee5b44ee
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2215228
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>