Inital CL to add in a Stackdriver nonitor to ts_mon.
This contains just an implementation of the Monitor interface using Stackdriver.
Following CLs will make the changes needed use this new Monitor as well as the
ability to use multiple monitors.

The primary use of this is to get Custom metrics accessible for Kayenta CAS.
Aim is also to make the ts-backend modular.

Change-Id: I2568f4efb719fbd895d58bd8d1815f55c26cb98e