[resultdb] Add artifacts package
Move artifact functionality from internal/span to internal/artifacts.
Note that names got shorter: e.g. "artifacts.ParentID" instead of
"span.ArtifactParentID".

In non-test code, the changes are mostly mechanical, or automated
renames. Most of manual changes are in test code.

This also enabled moving tests back to the package with code-under-test.
It was not possible before because it created a circular
dependency between span and testutil.

One notable change in non-test code is simplification of page
token parsing code, see the changes in the pagination package.

Bug: 1088278
Change-Id: I53a4b4bbb8569bb792cb3219eb4f3ae8a6affc8c