cache: use temp file while copying data from src
I'm completely not sure why the task like
https://chromium-swarm.appspot.com/task?id=4c8dab42bbb83310
shows 'permission denied' error.

But the error seems to happen for the files such that there are other
files having the same hash in an isolated. So I guess something strange
things happen when the client opens such files.
As a possible fix, let me use temp file and rename for atomic file
operations.

Bug: 1090293, 1076468
Change-Id: Icec6f679fdcc5e505b01e311b76760c6a77bc12b