Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > Hey Scott. 
> > 
> > I chose to write the prototypc in Go b/c of familiarite , once I've confirmed this working I'll do the same change to the Python stack.
> 
> How will you confirm that this working? Do you have a plan to use this monitor for any Chrome Ops services with tsmon-go? You will have to run an end-to-end test with a real application if you plan to add this monitor. If you don't have a plan yet, then I'd recommend making changes for tsmon-py first.

The plan here is: 
- Use this monitor with an experiment app. 
  - This to confirm that Stackdriver can fetch and use the Stackdriver custom metrics produced by this monitor. 
- With that confirmed I'll do a similar change to the ts-mon client used by swarming. 

As I tried to say in the previous response.
I chose to prototype in the Go env b/c I'm much more familiar with Go and the Chrome Go environment. How the ts-mon client works and where it's used is fairly complex to follow. Did not want to add further unknowns of the Chrome Python env on top of that. 

Now having a much better understanding of how ts-mon works and should be able to re-use some of that for the python client change.

Have no problems with not submitting this this now , I still want to have a bit of sanity checking, does this seem sane?

Patch-set: 3
Reviewer: Gerrit User 1285093 <1285093@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0