WIP: resolve external consoles before storing in datastore.
TODO: Put the required refactoring (getting all consoles in a map upfront) into
a separate CL.

The approach of this CL is as follows:
* When importing consoles from luci-config into Datastore, resolve any
external consoles such that we store the details of the external console
inside the Def proto.
* When rendering the console on the project page, we use the external
project and console IDs to create the console link.
* When rendering the console on the project page, we add an extra check
that the user can view the project that the console originally comes
from.. (NOT IMPLEMENTED YET)

Bug: 1068630
Change-Id: I4f89a7a76745f070662a34c589c6891fd52c2ab6